+++ NEWS.md
@@ -3,8 +3,12 @@
 ## Feature updates
 
 * Imported [`getTBinR::get_tb_burden()`](https://www.samabbott.co.uk/getTBinR/reference/get_tb_burden.html) in order to surface WHO TB burden estimates.
+* Added FAO estimates of population demographics to the package. 
+* Added OIE estimates for zoonostic TB presence in domesticated and wild animals.
+* Added a data sources vignette detailing available data, sources, and cleaning steps taken.
 
 ## Package updates
 
 * Set up package using existing infrastructure from [getTBinR](https://github.com/seabbs/getTBinR).
 * Documentation building set up using Travis CI.
+* Added tests for all datasets.